requirements.txt
fastapi
uvicorn
psycopg2-binary
boto3
pydantic-settings
torch